usb: xhci: Add support for SINGLE_STEP_SET_FEATURE test of EHSET
The Embedded High-speed Host Electrical Test (EHSET) procedure defines
the SINGLE_STEP_SET_FEATURE test for an embedded USB Host port. Upon
activating this test mode, the SETUP stage of a GetDescriptor request
is sent and followed by a delay of 15 seconds before finishing with
the DATA and STATUS stages. The idea is that this delay will give the
test operator sufficient time to configure the oscilloscope to perform
a measurement of the response time delay between the latter two stages.

This test is not implemented by the EHCI/xHCI host controller itself but
can be implemented in software. Similar to commit 75e845f64 ("usb: ehci:
Add support for SINGLE_STEP_SET_FEATURE test of EHSET"), this patch adds
such support to the xHCI driver.

CRs-fixed: 461710
Change-Id: I69aad8b068e40c13a58a66be917c185492615bfb
Signed-off-by: Jack Pham <jackp@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>