ASoC: wcd9xxx: Avoid plug type re-determination
In the case where there is mechanical interrupt
at the same time when polling scheduled for execution,
we should make sure that if plug type is found and reported
by the polling, re-determination from the insertion irq
should be avoided. Add a check in the insertion irq before
proceeding further to determine the plug type.

Change-Id: I0da017073700d3a362b6175f1292d603885b7ab1
Signed-off-by: Phani Kumar Uppalapati <phaniu@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>