msm: ADSPRPC: Always map buffers into IOMMU if one is present
If the buffers are not mapped into IOMMU and the physical address
is passed to remote processor, then it could result in IOMMU faults
if the passed physical address falls in the IOMMU virtual address
range. Map all buffers into IOMMU if one is present and do not pass
the physical address directly to the remote processor.

Change-Id: Id567ac4bff16f49487439d1d505dd25b88ed2868
Acked-by: Sathish Ambley <sambley@qualcomm.com>
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>